Improved UX for authentication when user changes his phone number
